requirements.txt
python-decouple==3.8
aiogram==3.4.1
